split up task for clarity
